[RFC] Set the 'default' choice to overwrite the .gitignore to 'no'.
When running `composer create-project`, all default choices result in default locations for the files and folders. The last question feels like it's the other way around: The default option overwrites the existing `.gitignore` with the one from `bolt/bolt`. 

I'd like 'n' to be the default, so that this is what it does when running the packaging script for the "end users".